Add distance (in km) column to Live Data frame

+++ src/za/co/turton/eyeburst/TowerDatum.java
@@ -29,7 +29,7 @@ public class TowerDatum {
     /**
      * The reported distance
      */
-    public int distance;
+    public float distance;
     
     /**
      * The reported load
